changed solution type to i64 + solved day9

+++ src/day11.rs
@@ -3,12 +3,11 @@ use crate::Solution;
 pub struct Day11;
 
 impl Solution for Day11 {
-
-    fn part1(input: &str) -> Option<usize> {
+    fn part1(input: &str) -> Option<i64> {
         None
     }
 
-    fn part2(input: &str) -> Option<usize> {
+    fn part2(input: &str) -> Option<i64> {
         None
     }
 }
