+++ src/day14.rs
@@ -3,12 +3,11 @@ use crate::Solution;
 pub struct Day14;
 
 impl Solution for Day14 {
-
-    fn part1(input: &str) -> Option<usize> {
+    fn part1(input: &str) -> Option<i64> {
         None
     }
 
-    fn part2(input: &str) -> Option<usize> {
+    fn part2(input: &str) -> Option<i64> {
         None
     }
 }
