+++ src/day21.rs
@@ -3,12 +3,11 @@ use crate::Solution;
 pub struct Day21;
 
 impl Solution for Day21 {
-
-    fn part1(input: &str) -> Option<usize> {
+    fn part1(input: &str) -> Option<i64> {
         None
     }
 
-    fn part2(input: &str) -> Option<usize> {
+    fn part2(input: &str) -> Option<i64> {
         None
     }
 }
